pipeline/requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
supabase==2.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1
